time buffer testbench
the current version does not test handling of the valid signal, but instead
prints out the time every time it changes. GHDL manages to simulate this
at a rate of a couple of simulated days per real minute at a clock
frequency of .5Hz, so I plan to gather at least a couple of decades worth
of time and compare that against a simple C program that prints the same
information.

Let's hope GHDL doesn't show any bugs if you run it for that long…

Signed-off-by: wrtlprnft <wrtlprnft@cssbook.de>